requirements.txt
beautifulsoup4==4.11.2
en-core-web-sm==3.5.0
faiss-cpu==1.7.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
matplotlib==3.5.1
mysql-connector==2.2.9
nltk==3.8.1
numpy==1.20.3
pandas==1.5.3
Pillow==8.4.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
spacy==3.5.0
SQLAlchemy==2.0.5